Capacity Workload and patient loss through Autorun and Mainform matched

+++ src/main/java/com/envision/Staffing/services/ShiftPlanningService.java
@@ -49,7 +49,7 @@ public class ShiftPlanningService {
 
 		input.setClinician(jobDetails.getClinicians().stream().toArray(Clinician[]::new));
 		input.setLowerLimitFactor(jobDetails.getLowerUtilizationFactor());
-		input.setUpperLimitFactor((double) jobDetails.getUpperUtilizationFactor());
+		input.setUpperLimitFactor(jobDetails.getUpperUtilizationFactor());
 		input.setNotAllocatedStartTime(jobDetails.getNotAllocatedStartTime());
 		input.setNotAllocatedEndTime(jobDetails.getNotAllocatedEndTime());
 		input.setPatientHourWait(jobDetails.getPatientHourWait());
